ai-core/requirements.txt
fastapi
uvicorn
pypdf
python-docx
regex
python-multipart
openai
google-generativeai
tenacity
python-dotenv
